seperate out lib-logic to ContractGenerator class from ProgramMain
